[Memory][Cartridges]: Write/Read access for MBC1-3 Cartridges
Differents gameboy cartridges have different mehtods of writing nd
acessing data.
MBC type cartridges include ROM memory and new registers.

We currently only support accessing and accessing data for MBC1, MBC2,
and MBC3 cartridges, which are the most common and, as such, should beb
sufficient for now. The other type of cartridges default to a "classic"
cartridge with no ROM when interfacing.

When trying to access or write to a cartridge we should always use the
`write_cartridge` or `write_memory` API as it will automatically detect
the type of the currently loaded cartridge and will proceed accordingly.

TODO:
- TESTING !!!
- MBC5-7 cartridges